FIX - Broken integration tests Fixed a bug in my test routine which meant that the Greenfield parser wasn't tested properly. - moved CorpusManager to src/main; and - fixed broken test method in ResultSetImpl.
