+++ src/test/java/org/verapdf/pdfa/validation/validators/ValidatorTest.java
@@ -19,7 +19,6 @@ import org.verapdf.core.EncryptedPdfException;
 import org.verapdf.core.ModelParsingException;
 import org.verapdf.core.ValidationException;
 import org.verapdf.core.XmlSerialiser;
-import org.verapdf.integration.CorpusManager;
 import org.verapdf.model.ModelParser;
 import org.verapdf.pdfa.Foundries;
 import org.verapdf.pdfa.PDFAParser;
@@ -27,6 +26,7 @@ import org.verapdf.pdfa.PDFAValidator;
 import org.verapdf.pdfa.PdfBoxFoundryProvider;
 import org.verapdf.pdfa.flavours.PDFAFlavour;
 import org.verapdf.pdfa.qa.AbstractTestCorpus.Corpus;
+import org.verapdf.pdfa.qa.CorpusManager;
 import org.verapdf.pdfa.qa.CorpusSampler;
 import org.verapdf.pdfa.qa.GitHubBackedProfileDirectory;
 import org.verapdf.pdfa.qa.TestCorpus;
